+++ changelog.md
@@ -48,3 +48,4 @@
 ## Next
 
 * `genWrap` option
+* Treat original functions which are generators as co-routines (breaking change)