+++ changelog.md
@@ -49,3 +49,4 @@
 
 * `genWrap` option
 * Treat original functions which are generators as co-routines (breaking change)
+* Code comments