+++ changelog.md
@@ -51,3 +51,4 @@
 * `genWrap` option
 * Treat original functions which are generators as co-routines (breaking change)
 * Code comments
+* README update